Make with_current method available in all tests
